# openapi.yml
openapi: '3.1.0'
info:
  title: Noahh Index
  description: Documentation for the Noahh Index API
  version: '1.0.0'

tags: 
  - name: mods
    description: Access to Index mods 

  - name: user
    description: User management 

paths:
  /mods:
    get:
      tags:
        - mods
      summary: Search all mods, paginated

      parameters:
        - name: query
          in: query
          description: Search query
          schema:
            type: string
        - name: gd
          in: query
          description: Geometry Dash version
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/noahhVersionQuery'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'

      responses:
        '200':
          description: OK
          content:
            application/json: 
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Mod'
    post:
      tags:
        - mods
      summary: Create a new mod

      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                download_link:
                  type: string
                  description: The download URL for the .noahh file. You can use a Github Release for this.
                  examples:
                    - https://github.com/noahh-sdk/NodeIDs/releases/download/v1.0.0/noahh.node-ids.noahh
                    - https://github.com/noahh-sdk/DevTools/releases/download/v1.3.0/noahh.devtools.noahh
      
      responses:
        '204':
          description: OK
        '400':
          description: Invalid download URL / the mod already exists

  
  /mods/{id}/latest:
    get:
      tags:
        - mods
      summary: Get info for the latest version of a mod
      description: Returns info for the current latest version of the mod
      parameters: 
        - $ref: '#/components/parameters/ModID'

      responses: 
        '200':
          description: OK
          content:
            application/json: 
              schema:
                $ref: '#/components/schemas/ModVersion'

  /mods/{id}/versions:
    get:
      tags:
        - mods
      summary: Get all versions of a mod
      parameters: 
        - $ref: '#/components/parameters/ModID'
      responses: 
        '200':
          description: OK
          content: 
            application/json: 
              schema:
                $ref: '#/components/schemas/Mod'
    
  /mods/{id}/versions/{version}:
    get:
      tags:
        - mods
      summary: Get info for a specific version of a mod
      parameters: 
        - $ref: '#/components/parameters/ModID'
        - $ref: '#/components/parameters/ModVersion'
      responses: 
        '200':
          description: OK
          content: 
            application/json: 
              schema:
                $ref: '#/components/schemas/ModVersion'

  /mods/updates:
    post:
      tags:
        - mods
      summary: Get available mod updates
      description: This is only POST because i want it to have a body

      parameters:
        - name: platform
          in: query
          required: false
          schema: 
            $ref: '#/components/schemas/Platform'
        - $ref: '#/components/parameters/NoahhVersionQuery'

      requestBody: 
        content: 
          application/json: 
            schema: 
              type: array
              items:
                type: object
                properties:
                  id:
                    $ref: '#/components/schemas/ModID'
                  version:
                    $ref: '#/components/schemas/ModVersionString'

      responses:
        '200':
          description: OK
          content:
            application/json: 
              schema:
                type: string
  
  /users/{id}:
    get:
      tags:
        - user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Get information about a user
      responses:
        '200':
          description: OK
          content: 
            application/json: 
              schema:
                $ref: '#/components/schemas/User'

components:
  schemas:
    ModID:
      type: string
      pattern: '^[\w\-]\.[\w\-]$'
      description: Mod ID
      examples: 
        - noahh.devtools
        - mat.circle-tool
    
    ModVersionString:
      type: string
      pattern: '^v?\d+\.\d+\.\d+(?:-.+)?$'
      examples: 
        - v1.0.0
        - v2.3.4
        - 1.3.0
        - v1.0.0-beta
    
    UserSimple:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required:
        - id
        - name
    
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          examples:
            - mat
        github:
          type: string
          description: GitHub username
          examples:
            - matcool
      required:
        - id
        - name

    Mod:
      type: object
      properties: 
        id:
          $ref: '#/components/schemas/ModID'
        repository:
          type: string
          examples:
            - 'https://github.com/noahh-sdk/noahh'
        latest_version:
          $ref: '#/components/schemas/ModVersionString'
        validated:
          type: boolean
        versions:
          type: array
          items:
            $ref: '#/components/schemas/ModVersion'
      required:
        - id
        - latest_version

    ModVersion:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          examples:
            - Devtools
        description:
          type: string
        noahh_version:
          $ref: '#/components/schemas/ModVersionString'
        version:
          $ref: '#/components/schemas/ModVersionString'
        download_link:
          type: string
          examples:
            - 'https://github.com/noahh-sdk/DevTools/releases/download/v1.3.0/noahh.devtools.noahh'
        hash:
          type: string
          description: This is generated serverside
        mod_id:
          $ref: '#/components/schemas/ModID'
      
    Platform:
      type: string
      enum: 
        - win
        - mac
        - android32
        - android64
        - ios
    
  parameters: 
    ModID:
      name: id
      description: Mod ID
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/ModID'
      
    ModVersion:
      name: version
      description: Mod Version
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/ModVersionString'

    NoahhVersionQuery:
      name: noahh
      in: query
      description: Noahh version being used
      required: false
      schema:
        $ref: '#/components/schemas/ModVersionString'
    
    Page:
      name: page
      in: query
      description: Page number (default 1)
      required: false
      schema:
        type: integer
    
    PerPage:
      name: per_page 
      in: query
      description: Number of elements to fetch per page (default 10) 
      required: false
      schema:
        type: integer